completed refactor alignment
